requirements.txt
numpy

pillow

tensorflow[and-cuda] 

scikit-learn

tqdm

matplotlib

datasets